Removed the separate postgres_query. Retained the separate test files so that column mapping can be tested.
